Displaying live bus data on the map

+++ app/src/main/java/app/busalert/model/Vehicle.java
@@ -4,6 +4,6 @@ public interface Vehicle {
 
     String getLine();
     String getBrigade();
-    long getLatitude();
-    long getLongitude();
+    double getLatitude();
+    double getLongitude();
 }
